Implement LocationsCreate lambda
This receives a POST request from the /locations/new route on the
frontend, and does the relevant GraphQL requests to the API to get the
provider id and then submit the site.

Some hacks are in place:

- The provider code is hard-coded (this can be extracted from the
  request)
- The error handling for the form body parsing is insufficient
- The `code` generation is incomplete and doesn't use all possible codes
- No validations are implemented
- Errors will display as json in the browser
- The redirect code is brittle

This can be refactored beautifully with some TDD, but is not
necessary for this proof of concept app.